Use integer subtraction.

+++ src/main/java/uk/ac/sussex/gdsc/core/utils/rng/NumberUtils.java
@@ -492,8 +492,8 @@ public final class NumberUtils {
   public static double makeSignedDouble(long bits) {
     // Upper 53-bits is a positive number in the range [0, 1).
     // This has 1 optionally subtracted. Do not use the lower bits on the
-    // assumption they are less random.
-    return ((bits >>> 11) * 0x1.0p-53d) - ((bits >>> 10) & 0x1);
+    // assumption they are less random. Use integer subtraction.
+    return ((bits >>> 11) * 0x1.0p-53d) - (int) ((bits >>> 10) & 0x1);
   }
 
   /**
